backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
python-dotenv
gunicorn
django-filter
drf-spectacular
djangorestframework-simplejwt
django-prometheus
